--- packages/client/src/entities/Player/index.ts
@@ -1,1 +1,0 @@
-export type { Player } from './Player';
