+++ packages/client/src/pages/game/ui/GamePage.module.scss
@@ -6,6 +6,7 @@
   display: flex;
   padding-top: 30px;
   justify-content: center;
+  box-sizing: border-box;
 
   .game {
     display: flex;
@@ -17,6 +18,7 @@
       @include typography.font(h2);
 
       color: var(--color-yellow-text);
+      text-shadow: var(--shadow-yellow-text);
       font-size: 5rem;
     }
   }
