+++ packages/client/src/shared/const/game.ts
@@ -1,0 +1,1 @@
+export const tileSize = 50;
