+++ packages/client/src/shared/const/ghostImages.ts
@@ -1,0 +1,6 @@
+const colors = ['pink', 'green', 'blue'];
+export const ghostImages: HTMLImageElement[] = colors.map((color) => {
+  const img = new Image();
+  img.src = `/src/assets/images/ghosts/${color}.svg`;
+  return img;
+});
