+++ packages/client/src/shared/model/direction.ts
@@ -1,0 +1,10 @@
+import { Vector2D } from './vector';
+
+export type Direction = 'ArrowUp' | 'ArrowDown' | 'ArrowLeft' | 'ArrowRight';
+
+export const directionVectors: Record<Direction, Vector2D> = {
+  ArrowUp: { x: 0, y: -1 },
+  ArrowDown: { x: 0, y: 1 },
+  ArrowLeft: { x: -1, y: 0 },
+  ArrowRight: { x: 1, y: 0 }
+};
