+++ packages/client/src/shared/ui/index.ts
@@ -1,3 +1,2 @@
 export { Button } from './Button/Button';
 export { Input } from './Input/Input';
-export { Game } from './Game/Game';
