+++ packages/client/src/widgets/GameBoard/index.ts
@@ -1,0 +1,1 @@
+export { GameBoard } from './ui/GameBoard';
