+++ packages/client/src/widgets/GameBoard/ui/GameBoard.module.scss
@@ -4,4 +4,5 @@
   @include typography.font(h2);
 
   color: var(--color-yellow-text);
+  text-shadow: var(--shadow-yellow-text);
 }
